Feat (graph_quant): new quantize functions and modules
